pyproject.toml
[tool.poetry]
name = "world-spending-perspective"
version = "0.1.0"
description = "COFOG drill-down ETL infra scaffold"
authors = ["Your Name <your@email.com>"]

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
pydantic = "^2.7.0"
duckdb = "^0.10.0"
pyarrow = "^16.0.0"
pytest = "^8.2.0"
rich = "^13.7.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
